+++ doc/tut-3-recv.rst
@@ -2,8 +2,8 @@ Receiver, version 1
 ===================
 Now that we have a sender, let's write a receiver. For the sake of an example,
 let's have it report the average power (mean squared value) of the samples in
-each heap. The full code can be found in :file:`examples/tut_recv_1.py` and
-:file:`examples/tut_recv_2.cpp` in the spead2 repository.
+each heap. The full code can be found in :file:`examples/tut_3_recv.py` and
+:file:`examples/tut_3_recv.cpp` in the spead2 repository.
 
 The initial boilerplate looks similar to the sender, and once again we'll need
 a thread pool.
