+++ examples/tut_2_send.cpp
@@ -19,9 +19,6 @@
 #include <string>
 #include <vector>
 #include <utility>
-#include <chrono>
-#include <memory>
-#include <iostream>
 #include <boost/asio.hpp>
 #include <spead2/common_defines.h>
 #include <spead2/common_thread_pool.h>
@@ -29,26 +26,18 @@
 #include <spead2/send_stream_config.h>
 #include <spead2/send_udp.h>
 
-struct state
-{
-    std::future<spead2::item_pointer_t> future;
-    std::vector<std::int8_t> adc_samples;
-    spead2::send::heap heap;
-};
-
 int main()
 {
     spead2::thread_pool thread_pool;
     spead2::send::stream_config config;
-    config.set_rate(0.0);
-    config.set_max_heaps(2);
+    config.set_rate(100e6);
     boost::asio::ip::udp::endpoint endpoint(
         boost::asio::ip::address::from_string("127.0.0.1"),
         8888
     );
     spead2::send::udp_stream stream(thread_pool, {endpoint}, config);
 
-    const std::int64_t chunk_size = 1024 * 1024;
+    const std::int64_t heap_size = 1024 * 1024;
     spead2::descriptor timestamp_desc;
     timestamp_desc.id = 0x1600;
     timestamp_desc.name = "timestamp";
@@ -59,20 +48,15 @@ int main()
     adc_samples_desc.name = "adc_samples";
     adc_samples_desc.description = "ADC converter output";
     adc_samples_desc.numpy_header =
-        "{'shape': (" + std::to_string(chunk_size) + ",), 'fortran_order': False, 'descr': 'i1'}";
+        "{'shape': (" + std::to_string(heap_size) + ",), 'fortran_order': False, 'descr': 'i1'}";
 
     std::default_random_engine random_engine;
     std::uniform_int_distribution<std::int8_t> distribution(-100, 100);
+    std::vector<std::int8_t> adc_samples(heap_size);
 
-    const int n_heaps = 100;
-    auto start = std::chrono::high_resolution_clock::now();
-    std::unique_ptr<state> old_state;
-    for (int i = 0; i < n_heaps; i++)
+    for (int i = 0; i < 10; i++)
     {
-        auto new_state = std::make_unique<state>();
-        auto &heap = new_state->heap;
-        auto &adc_samples = new_state->adc_samples;
-        adc_samples.resize(chunk_size);
+        spead2::send::heap heap;
         // Add descriptors to the first heap
         if (i == 0)
         {
@@ -80,25 +64,18 @@ int main()
             heap.add_descriptor(adc_samples_desc);
         }
         // Create random data
-        for (int j = 0; j < chunk_size; j++)
+        for (int j = 0; j < heap_size; j++)
             adc_samples[j] = distribution(random_engine);
         // Add the data and timestamp to the heap
-        heap.add_item(timestamp_desc.id, i * chunk_size);
+        heap.add_item(timestamp_desc.id, i * heap_size);
         heap.add_item(
             adc_samples_desc.id,
             adc_samples.data(),
             adc_samples.size() * sizeof(adc_samples[0]),
             true
         );
-        new_state->future = stream.async_send_heap(heap, boost::asio::use_future);
-        if (old_state)
-            old_state->future.get();
-        old_state = std::move(new_state);
+        stream.async_send_heap(heap, boost::asio::use_future).get();
     }
-    old_state->future.get();
-    auto elapsed = std::chrono::duration_cast<std::chrono::duration<double>>(
-        std::chrono::high_resolution_clock::now() - start);
-    std::cout << chunk_size * n_heaps / elapsed.count() / 1e6 << " MB/s\n";
 
     // Send an end-of-stream control item
     spead2::send::heap heap;
