+++ examples/tut_4_send_pipeline.cpp
@@ -15,6 +15,7 @@
  */
 
 #include <cstdint>
+#include <random>
 #include <string>
 #include <vector>
 #include <utility>
@@ -35,26 +36,19 @@ struct state
     spead2::send::heap heap;
 };
 
-int main(int argc, char * const argv[])
+int main()
 {
-    if (argc != 3)
-    {
-        std::cerr << "Usage: " << argv[0] << " <address> <port>\n";
-        return 2;
-    }
-
     spead2::thread_pool thread_pool;
     spead2::send::stream_config config;
     config.set_rate(0.0);
     config.set_max_heaps(2);
-    config.set_max_packet_size(9000);
     boost::asio::ip::udp::endpoint endpoint(
-        boost::asio::ip::address::from_string(argv[1]),
-        std::atoi(argv[2])
+        boost::asio::ip::address::from_string("127.0.0.1"),
+        8888
     );
     spead2::send::udp_stream stream(thread_pool, {endpoint}, config);
 
-    const std::int64_t chunk_size = 1024 * 1024;
+    const std::int64_t heap_size = 1024 * 1024;
     spead2::descriptor timestamp_desc;
     timestamp_desc.id = 0x1600;
     timestamp_desc.name = "timestamp";
@@ -65,9 +59,12 @@ int main(int argc, char * const argv[])
     adc_samples_desc.name = "adc_samples";
     adc_samples_desc.description = "ADC converter output";
     adc_samples_desc.numpy_header =
-        "{'shape': (" + std::to_string(chunk_size) + ",), 'fortran_order': False, 'descr': 'i1'}";
+        "{'shape': (" + std::to_string(heap_size) + ",), 'fortran_order': False, 'descr': 'i1'}";
 
-    const int n_heaps = 10000;
+    std::default_random_engine random_engine;
+    std::uniform_int_distribution<std::int8_t> distribution(-100, 100);
+
+    const int n_heaps = 100;
     auto start = std::chrono::high_resolution_clock::now();
     std::unique_ptr<state> old_state;
     for (int i = 0; i < n_heaps; i++)
@@ -75,16 +72,18 @@ int main(int argc, char * const argv[])
         auto new_state = std::make_unique<state>();
         auto &heap = new_state->heap;
         auto &adc_samples = new_state->adc_samples;
-        // Fill with the heap number
-        adc_samples.resize(chunk_size, i);
+        adc_samples.resize(heap_size);
         // Add descriptors to the first heap
         if (i == 0)
         {
             heap.add_descriptor(timestamp_desc);
             heap.add_descriptor(adc_samples_desc);
         }
+        // Create random data
+        for (int j = 0; j < heap_size; j++)
+            adc_samples[j] = distribution(random_engine);
         // Add the data and timestamp to the heap
-        heap.add_item(timestamp_desc.id, i * chunk_size);
+        heap.add_item(timestamp_desc.id, i * heap_size);
         heap.add_item(
             adc_samples_desc.id,
             adc_samples.data(),
@@ -99,7 +98,7 @@ int main(int argc, char * const argv[])
     old_state->future.get();
     auto elapsed = std::chrono::duration_cast<std::chrono::duration<double>>(
         std::chrono::high_resolution_clock::now() - start);
-    std::cout << chunk_size * n_heaps / elapsed.count() / 1e6 << " MB/s\n";
+    std::cout << heap_size * n_heaps / elapsed.count() / 1e6 << " MB/s\n";
 
     // Send an end-of-stream control item
     spead2::send::heap heap;
