+++ examples/tut_6_recv_power.py
